changed ray_chastetris.h to chastetris.h
The header of logic for my game has had "ray" in it's name removed because it contained absolutely nothing raylib related. I proved this by using the exact same file for my SDL port. It's pure ansi C now. The font library is a different story and is platform specific because it has to load and draw images using the current media library.